take out localization loading into seperate file
